changed type of averageWeight from double to int removed unnecessary imports in service

+++ src/main/java/org/marcinski/chickenHouse/dto/DayDto.java
@@ -24,7 +24,7 @@ public class DayDto {
     private long waterCounter;
 
     @Nullable
-    private double averageWeight;
+    private int averageWeight;
 
     @Nullable
     private String comments;
